requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
google-genai==1.70.0
boto3

